refactor: split AdminPage into per-section components
AdminPage.tsx was a single 290-line file containing the layout, four
section components, the fetch hook, formatters, the table component,
the section wrapper, and a status-line component. Split into focused
files under packages/client/src/features/admin/:

- formatters.ts — formatTimestamp + formatDuration. The latter now
  uses Intl.RelativeTimeFormat for locale-aware "in 5 minutes" /
  "5 minutes ago" output instead of hand-rolled "5m" abbreviations.
- useFetchOnce.ts — the mount-once data fetch hook + FetchState type.
- Section.tsx — Section + StatusLine helpers used by every section.
- DataTable.tsx — generic table with column-config pattern.
- UsersSection.tsx, SessionsSection.tsx, KeepaliveSection.tsx,
  LogsSection.tsx — one section per file.

AdminPage.tsx now just composes the four sections.
