+++ packages/client/src/features/admin/DataTable.tsx
@@ -1,0 +1,37 @@
+import type { ReactNode } from 'react';
+
+export interface Column<T> {
+  header: string;
+  cell: (row: T) => ReactNode;
+}
+
+export function DataTable<T>({
+  rows,
+  columns,
+  getKey,
+}: {
+  rows: T[];
+  columns: Column<T>[];
+  getKey: (row: T, index: number) => string | number;
+}) {
+  return (
+    <table className="admin-table">
+      <thead>
+        <tr>
+          {columns.map((c) => (
+            <th key={c.header}>{c.header}</th>
+          ))}
+        </tr>
+      </thead>
+      <tbody>
+        {rows.map((row, i) => (
+          <tr key={getKey(row, i)}>
+            {columns.map((c) => (
+              <td key={c.header}>{c.cell(row)}</td>
+            ))}
+          </tr>
+        ))}
+      </tbody>
+    </table>
+  );
+}
